Alter Documentation link to always point to current
Rather than constantly bumping the documentation link when a new
release is out we can use http://gpdb.docs.pivotal.io/ which will
redirect to the current version. This will have to be revisited
once we get the documentation on greenplum.org but for now, let's
make sure we are in sync.

Reported by Mel Kiyama